+++ src/store/index.js
@@ -17,4 +17,6 @@ export const store = configureStore({
 
 // optional, but required for refetchOnFocus/refetchOnReconnect behaviors
 // see `setupListeners` docs - takes an optional callback as the 2nd arg for customization
+
+
 setupListeners(store.dispatch);